+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 ## Next
 
-- `flush` forcefully sends events regardless the delay ([#66](https://github.com/PostHog/posthog-android/pull/66))
+- `flush` forcefully sends events regardless the delay ([#73](https://github.com/PostHog/posthog-android/pull/73))
 
 ## 3.0.0 - 2023-12-06